+++ resources/views/admin/users/_form.blade.php
@@ -1,38 +1,41 @@
-<div>
-    <div class="row-start-start gutter-4">
-        <x-chief-form::formgroup id="firstname" label="Voornaam" class="w-full lg:w-1/2" required>
-            <input id="firstname" type="text" name="firstname" value="{{ old('firstname', $user->firstname) }}">
-            <x-chief-form::formgroup.error error-ids="firstname"/>
-        </x-chief-form::formgroup>
+<div class="row-start-start gutter-3">
+    <x-chief::input.group rule="firstname" class="w-full lg:w-1/2">
+        <x-chief::input.label for="firstname" required>Voornaam</x-chief::input.label>
+        <x-chief::input.text id="firstname" name="firstname" value="{{ old('firstname', $user->firstname) }}"/>
+    </x-chief::input.group>
 
-        <x-chief-form::formgroup id="lastname" label="Achternaam" class="w-full lg:w-1/2" required>
-            <input id="lastname" type="text" name="lastname" value="{{ old('lastname', $user->lastname) }}">
-            <x-chief-form::formgroup.error error-ids="lastname"/>
-        </x-chief-form::formgroup>
-    </div>
+    <x-chief::input.group rule="lastname" class="w-full lg:w-1/2">
+        <x-chief::input.label for="lastname" required>Achternaam</x-chief::input.label>
+        <x-chief::input.text id="lastname" name="lastname" value="{{ old('lastname', $user->lastname) }}"/>
+    </x-chief::input.group>
+
+    <x-chief::input.group rule="email" class="w-full">
+        <x-chief::input.label for="email" required>
+            E-mailadres
+        </x-chief::input.label>
+
+        <x-chief::input.description>
+            Dit e-mailadres geldt tevens als login.
+        </x-chief::input.description>
+
+        <x-chief::input.email id="email" name="email" value="{{ old('email', $user->email) }}"/>
+    </x-chief::input.group>
+
+    <x-chief::input.group rule="roles" class="w-full">
+        <x-chief::input.label for="roles" required>
+            Rechten
+        </x-chief::input.label>
+
+        <x-chief::input.description>
+            Geef aan met een of meerdere rollen welke rechten deze gebruiker ter beschikking heeft.
+        </x-chief::input.description>
+
+        <chief-multiselect
+            id="roles"
+            name="roles"
+            :options=@json($roleNames)
+            selected='@json(old('roles', $user->roleNames()))'
+            :multiple="true"
+        />
+    </x-chief::input.group>
 </div>
-
-<x-chief-form::formgroup id="email" label="E-mail (tevens login)" required>
-    <x-slot name="description">
-        Dit e-mail adres geldt tevens als login.
-    </x-slot>
-
-    <input id="email" type="email" name="email" value="{{ old('email', $user->email) }}">
-
-    <x-chief-form::formgroup.error error-ids="email"/>
-</x-chief-form::formgroup>
-
-<x-chief-form::formgroup id="roles" label="Rechten" required>
-    <x-slot name="description">
-        Geef aan met een of meerdere rollen welke rechten deze gebruiker ter beschikking heeft.
-    </x-slot>
-
-    <chief-multiselect
-        name="roles"
-        :options=@json($roleNames)
-        selected='@json(old('roles', $user->roleNames()))'
-        :multiple="true"
-    />
-
-    <x-chief-form::formgroup.error error-ids="roles"/>
-</x-chief-form::formgroup>