+++ resources/views/admin/users/create.blade.php
@@ -13,9 +13,7 @@
         <form id="createForm" action="{{ route('chief.back.users.store') }}" method="POST" class="card">
             @csrf
 
-            <div class="space-y-6">
-                @include('chief::admin.users._form')
-            </div>
+            @include('chief::admin.users._form')
         </form>
     </x-chief::page.grid>
 </x-chief::page.template>