+++ resources/views/manager/filters/radio.blade.php
@@ -1,22 +1,31 @@
-<x-chief-form::formgroup id="{{ $id }}" label="{{ $label ?? null }}">
-    @if(isset($description))
-        <x-slot name="description">
-            <p>{!! $description !!}</p>
-        </x-slot>
+@php
+    $label = $label ?? null;
+    $description = $description ?? null;
+@endphp
+
+<x-chief::input.group :rule="$id" inner-class="space-y-2">
+    @if ($label)
+        <x-chief::input.label>{{ $label }}</x-chief::input.label>
+    @endif
+
+    @if ($description)
+        <x-chief::input.description>{{ $description }}</x-chief::input.description>
     @endif
 
     <div class="space-y-1">
         @foreach($options as $option => $optionLabel)
-            <label class="with-radio" for="{{ $id.'-'.$option }}">
-                <input
-                    {{ ($option == ($value ?: $default)) ? 'checked="checked"':'' }}
+            <div class="flex items-start gap-2">
+                <x-chief::input.radio
+                    id="{{ $id . '-' . $option }}"
                     name="{{ $name }}"
                     value="{{ $option }}"
-                    id="{{ $id.'-'.$option }}"
-                    type="radio"
-                >
-                <span>{!! $optionLabel !!}</span>
-            </label>
+                    :checked="($option == ($value ?: $default))"
+                />
+
+                <x-chief::input.label for="{{ $id . '-' . $option }}">
+                    {{ $optionLabel }}
+                </x-chief::input.label>
+            </div>
         @endforeach
     </div>
-</x-chief-form::formgroup>
+</x-chief::input.group>